split analog readers to a separate file
